kites/terraformer: inject terraformer into kite
